# src/main/java/co/golmanager/gestorweb/presentation/dto/authentication/AuthenticationRequest.java
package co.golmanager.gestorweb.presentation.dto.authentication;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthenticationRequest {
    private String email;
    private String password;
}