# src/main/java/co/golmanager/gestorweb/presentation/dto/authentication/RegisterRequest.java
package co.golmanager.gestorweb.presentation.dto.authentication;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RegisterRequest {
    private String name;
    private String email;
    private String password;
}