# src/main/java/co/golmanager/gestorweb/presentation/dto/team/CreateTeamResponse.java
package co.golmanager.gestorweb.presentation.dto.team;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CreateTeamResponse {
    String message;
}