# src/main/resources/static/docs/openapi.json
{"openapi":"3.0.1","info":{"title":"GolManager API","description":"API for managing soccer tournaments, teams, players, and referees.","contact":{"name":"Mateo Vargas Tirado - Luis Carlos Vanegas Zapata"},"license":{"name":" AGPL-3.0 license"},"version":"0.2-SNAPSHOT"},"servers":[{"url":"http://localhost:8085","description":"Local server"}],"tags":[{"name":"Tournaments","description":"Endpoints for managing tournaments"},{"name":"Referees","description":"Endpoints for managing referees"},{"name":"Authentication","description":"Controller of Auth"},{"name":"Teams","description":"Endpoints for managing teams."},{"name":"Tournaments","description":"Endpoints for managing matches"}],"paths":{"/api/tournaments/{idTournament}/teams/{idTeam}":{"get":{"tags":["Teams"],"operationId":"getTeamDetails","parameters":[{"name":"idTournament","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idTeam","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamDetailsResponse"}}}}},"security":[{"Bearer Authentication":[]}]},"put":{"tags":["Teams"],"summary":"Update an existing tournament","description":"Update the details of existing tournament identified by its ID.","operationId":"updateTeamInfo","parameters":[{"name":"idTeam","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idTournament","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTeamRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamDetailsResponse"}}}}},"security":[{"Bearer Authentication":[]}]},"delete":{"tags":["Teams"],"summary":"Delete an team","description":"Delete team (players, score, goals, cards) of existing tournament","operationId":"deleteTeam","parameters":[{"name":"idTournament","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idTeam","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GeneralDeleteResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/tournaments/{Id}":{"get":{"tags":["Tournaments"],"summary":"Get tournament details by ID","description":"Retrieves detailed information about a specific tournament by its ID.","operationId":"getTournamentById","parameters":[{"name":"Id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentDetailResponse"}}}}},"security":[{"Bearer Authentication":[]}]},"put":{"tags":["Tournaments"],"summary":"Update an existing tournament","description":"Updates the details of an existing tournament identified by its ID.","operationId":"updateTournament","parameters":[{"name":"Id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTournamentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentDetailResponse"}}}}},"security":[{"Bearer Authentication":[]}]},"delete":{"tags":["Tournaments"],"summary":"Delete a tournament","description":"Deletes a tournament identified by its ID.","operationId":"deleteTournament","parameters":[{"name":"Id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TournamentDeleteResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/tournaments/{tournamentId}/matches":{"post":{"tags":["Tournaments"],"summary":"Create a match with date","operationId":"createMatch","parameters":[{"name":"tournamentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateMatchResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/tournaments/{idTournament}/teams":{"get":{"tags":["Teams"],"summary":"Obtain all teams for tournament","operationId":"getAllTeamsByTournament","parameters":[{"name":"idTournament","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetTeamsTournamentSummaryResponse"}}}}}},"security":[{"Bearer Authentication":[]}]},"post":{"tags":["Teams"],"summary":"Create a team for selected tournament","operationId":"createTeam","parameters":[{"name":"idTournament","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTeamRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateTeamResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/tournaments/create":{"post":{"tags":["Tournaments"],"summary":"Create a new tournament","description":"Creates a new tournament with the provided details.","operationId":"createTournament","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTournamentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateTournamentResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Authenticate","Authentication"],"summary":"Login User","description":"Authenticate a user and return the authentication token along wit user details.","operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"},"examples":{"Successful Login":{"description":"Successful Login","value":{"user":"organizador@torneos.com","password":"organizador123"}}}}},"required":true},"responses":{"200":{"description":"Succesful authentication","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"},"examples":{"Successful Login":{"description":"Successful Login","value":{"Token":"eyJhbGciOiJIUzI1NiJ9.eyJyb2xlIjoiVVNFUiIsInN1YiI6Im9yZ2FuaXphZG9yQHRvcm5lb3MuY29tIiwiaWF0IjoxNzU4OTMxMzk0LCJleHAiOjE3NTg5Mzg1OTR9.9EfE3GP_ZfVFDHIaxpbmVAS3_NajHdJySi8kSRe4dpA"}}}}}}}}},"/api/tournaments":{"get":{"tags":["Tournaments"],"summary":"Obtain all tournaments for the authenticated user","description":"Retrieves a list of all tournaments associated with the authenticated user.","operationId":"getAllTournamentsById","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TournamentSummaryResponse"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/referees":{"get":{"tags":["Referees"],"operationId":"listReferees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RefereeListResponse"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/greeting/time":{"get":{"tags":["greeting-controller"],"operationId":"time","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/greeting/sayHelloPublic":{"get":{"tags":["greeting-controller"],"operationId":"SayHello","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/greeting/sayHelloProtected":{"get":{"tags":["greeting-controller"],"operationId":"SayHelloProtected","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateTeamRequest":{"required":["coach","mainStadium","name","secondaryStadium","teamCategory"],"type":"object","properties":{"name":{"type":"string"},"coach":{"type":"string"},"teamCategory":{"type":"string","enum":["SUB13","SUB14","SUB15","SUB16","SUB17","SUB19","SUB20","SUB21","PROFESSIONAL","VETERAN"]},"mainStadium":{"type":"string"},"secondaryStadium":{"type":"string"}}},"TeamDetailsResponse":{"type":"object","properties":{"teamId":{"type":"integer","format":"int64"},"name":{"type":"string"},"coach":{"type":"string"},"category":{"type":"string","enum":["SUB13","SUB14","SUB15","SUB16","SUB17","SUB19","SUB20","SUB21","PROFESSIONAL","VETERAN"]},"mainStadium":{"type":"string"},"secondaryStadium":{"type":"string"},"dateCreated":{"type":"string","format":"date-time"}}},"CreateTournamentRequest":{"required":["endDate","format","homeAndAway","name","numberOfTeams","startDate","yellowCardsSuspension"],"type":"object","properties":{"name":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"format":{"type":"string","enum":["LEAGUE","DIRECT_ELIMINATION","PLAY_OFF"]},"homeAndAway":{"type":"boolean"},"numberOfTeams":{"type":"integer","format":"int32"},"yellowCardsSuspension":{"type":"integer","format":"int32"},"refereeIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"TournamentDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"format":{"type":"string","enum":["LEAGUE","DIRECT_ELIMINATION","PLAY_OFF"]},"homeAndAway":{"type":"boolean"},"numberOfTeams":{"type":"integer","format":"int32"},"yellowCardsSuspension":{"type":"integer","format":"int32"},"refereeIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"CreateMatchRequest":{"type":"object","properties":{"homeTeamId":{"type":"integer","format":"int64"},"awayTeamId":{"type":"integer","format":"int64"},"tournamentId":{"type":"integer","format":"int64"},"stadiumName":{"type":"string"},"referee":{"type":"integer","format":"int64"},"matchDate":{"type":"string","format":"date"}}},"CreateMatchResponse":{"type":"object","properties":{"matchId":{"type":"integer","format":"int32"},"homeTeam":{"type":"string"},"awayTeam":{"type":"string"},"stadiumName":{"type":"string"},"matchDate":{"type":"string","format":"date-time"}}},"CreatePlayerRequest":{"required":["age","name","playerPosition","shirtNumber"],"type":"object","properties":{"name":{"type":"string"},"age":{"type":"integer","format":"int32"},"playerPosition":{"type":"string","enum":["PO","DFI","DFC","DFD","MCD","MC","MCO","EI","DC","ED"]},"shirtNumber":{"type":"integer","format":"int32"}}},"CreateTeamRequest":{"required":["mainStadium","name","secondaryStadium","teamCategory","teamPlayers"],"type":"object","properties":{"name":{"type":"string"},"coach":{"type":"string"},"teamCategory":{"type":"string","enum":["SUB13","SUB14","SUB15","SUB16","SUB17","SUB19","SUB20","SUB21","PROFESSIONAL","VETERAN"]},"mainStadium":{"type":"string"},"secondaryStadium":{"type":"string"},"teamPlayers":{"type":"array","items":{"$ref":"#/components/schemas/CreatePlayerRequest"}}}},"CreateTeamResponse":{"type":"object","properties":{"message":{"type":"string"}}},"CreateTournamentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"format":{"type":"string"},"homeAndAway":{"type":"boolean"},"numberOfTeams":{"type":"integer","format":"int32"},"yellowCardsSuspension":{"type":"integer","format":"int32"},"refereeIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"}}},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"TournamentSummaryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"numberOfTeams":{"type":"integer","format":"int32"},"format":{"type":"string","enum":["LEAGUE","DIRECT_ELIMINATION","PLAY_OFF"]}}},"GetTeamsTournamentSummaryResponse":{"type":"object","properties":{"teamName":{"type":"string"},"coachName":{"type":"string"},"mainStadium":{"type":"string"},"teamCategory":{"type":"string","enum":["SUB13","SUB14","SUB15","SUB16","SUB17","SUB19","SUB20","SUB21","PROFESSIONAL","VETERAN"]}}},"Referee":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"RefereeListResponse":{"type":"object","properties":{"referees":{"type":"array","items":{"$ref":"#/components/schemas/Referee"}}}},"GeneralDeleteResponse":{"type":"object","properties":{"elementId":{"type":"integer","format":"int64"},"elementName":{"type":"string"},"deletionElementDate":{"type":"string","format":"date-time"}}},"TournamentDeleteResponse":{"type":"object","properties":{"tournamentId":{"type":"integer","format":"int64"},"name":{"type":"string"},"deletionDate":{"type":"string","format":"date-time"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}